Bot/requirements.txt
aiomysql>=0.1.1
requests>=2.28.1
aiohttp>=3.8.1
python-dotenv>=0.21.0